relevancy_tagger/requirements.txt
streamlit==1.32.0
pandas==2.2.0
pyyaml==6.0.1
SQLAlchemy==2.0.27
requests==2.31.0
python-dotenv==1.0.1
